backend/requirements.txt
fastapi
uvicorn
requests
Pillow
opencv-python
numpy
scikit-fuzzy
scipy
python-dotenv
httpx
